🐛 fix: reservation 통신 오류 해결
